# api/gen/models.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package gen

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for CreateUserRequestRole.
const (
	CreateUserRequestRoleAdmin  CreateUserRequestRole = "admin"
	CreateUserRequestRoleViewer CreateUserRequestRole = "viewer"
)

// Defines values for SatelliteManagedBy.
const (
	Manual   SatelliteManagedBy = "manual"
	Operator SatelliteManagedBy = "operator"
)

// Defines values for SatelliteStatus.
const (
	Error       SatelliteStatus = "Error"
	Pending     SatelliteStatus = "Pending"
	Ready       SatelliteStatus = "Ready"
	Unreachable SatelliteStatus = "Unreachable"
)

// Defines values for UserRole.
const (
	UserRoleAdmin  UserRole = "admin"
	UserRoleViewer UserRole = "viewer"
)

// CreateSatelliteRequest defines model for CreateSatelliteRequest.
type CreateSatelliteRequest struct {
	Name   string `json:"name"`
	Region string `json:"region"`
}

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	Password string                 `json:"password"`
	Role     *CreateUserRequestRole `json:"role,omitempty"`
	Username string                 `json:"username"`
}

// CreateUserRequestRole defines model for CreateUserRequest.Role.
type CreateUserRequestRole string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error *struct {
		Code    *string `json:"code,omitempty"`
		Message *string `json:"message,omitempty"`
	} `json:"error,omitempty"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Db     *string `json:"db,omitempty"`
	Status *string `json:"status,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	ExpiresIn *int    `json:"expiresIn,omitempty"`
	Token     *string `json:"token,omitempty"`
}

// Satellite defines model for Satellite.
type Satellite struct {
	CreatedAt  *time.Time          `json:"created_at,omitempty"`
	Id         *openapi_types.UUID `json:"id,omitempty"`
	LastSeenAt *time.Time          `json:"last_seen_at"`
	ManagedBy  *SatelliteManagedBy `json:"managed_by,omitempty"`
	Name       *string             `json:"name,omitempty"`
	Region     *string             `json:"region,omitempty"`
	Status     *SatelliteStatus    `json:"status,omitempty"`
}

// SatelliteManagedBy defines model for Satellite.ManagedBy.
type SatelliteManagedBy string

// SatelliteStatus defines model for Satellite.Status.
type SatelliteStatus string

// UpdateSatelliteRequest defines model for UpdateSatelliteRequest.
type UpdateSatelliteRequest struct {
	Region string `json:"region"`
}

// User defines model for User.
type User struct {
	CreatedAt *time.Time          `json:"created_at,omitempty"`
	Id        *openapi_types.UUID `json:"id,omitempty"`
	Role      *UserRole           `json:"role,omitempty"`
	Username  *string             `json:"username,omitempty"`
}

// UserRole defines model for User.Role.
type UserRole string

// VersionResponse defines model for VersionResponse.
type VersionResponse struct {
	Version *string `json:"version,omitempty"`
}

// Forbidden defines model for Forbidden.
type Forbidden = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorResponse

// PostAuthLoginJSONRequestBody defines body for PostAuthLogin for application/json ContentType.
type PostAuthLoginJSONRequestBody = LoginRequest

// PostSatellitesJSONRequestBody defines body for PostSatellites for application/json ContentType.
type PostSatellitesJSONRequestBody = CreateSatelliteRequest

// PatchSatellitesIdJSONRequestBody defines body for PatchSatellitesId for application/json ContentType.
type PatchSatellitesIdJSONRequestBody = UpdateSatelliteRequest

// PostUsersJSONRequestBody defines body for PostUsers for application/json ContentType.
type PostUsersJSONRequestBody = CreateUserRequest