# api/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: FleetControl API
  version: "0.1.0"
  description: Control Plane API for GitOps-based edge fleet management

servers:
  - url: http://localhost:8080

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Satellite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        region:
          type: string
        status:
          type: string
          enum: [Pending, Ready, Error, Unreachable]
        managed_by:
          type: string
          enum: [manual, operator]
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    CreateSatelliteRequest:
      type: object
      required: [name, region]
      properties:
        name:
          type: string
        region:
          type: string

    UpdateSatelliteRequest:
      type: object
      required: [region]
      properties:
        region:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [admin, viewer]
        created_at:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, viewer]
          default: viewer

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: integer

    HealthResponse:
      type: object
      properties:
        status:
          type: string
        db:
          type: string

    VersionResponse:
      type: object
      properties:
        version:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - BearerAuth: []

paths:
  /auth/login:
    post:
      summary: Login and get JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /satellites:
    get:
      summary: List all satellites
      responses:
        '200':
          description: List of satellites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Satellite'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a satellite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSatelliteRequest'
      responses:
        '201':
          description: Satellite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /satellites/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a satellite by ID
      responses:
        '200':
          description: Satellite found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      summary: Update a satellite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSatelliteRequest'
      responses:
        '200':
          description: Satellite updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a satellite
      responses:
        '204':
          description: Satellite deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /satellites/{id}/heartbeat:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Agent heartbeat
      responses:
        '200':
          description: Heartbeat received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
        '404':
          $ref: '#/components/responses/NotFound'

  /users:
    get:
      summary: List all users (admin only)
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'

  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Delete a user (admin only)
      responses:
        '204':
          description: User deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /health:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /version:
    get:
      summary: Version info
      security: []
      responses:
        '200':
          description: Version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'

 